refactoring/ssim: Split SSIM and MS-SSIM into separate files. Close #203
